save 3

+++ src/Add.php
@@ -5,7 +5,7 @@ namespace bsa1;
 
 class Add implements Operation
 {
-    public function evaluate($operands) : float
+    public function evaluate($operands) : int
     {
         $res = 0;
         foreach ($operands as $num) {
